TODO DebeziumJsonFormatter
